// admin/src/modulos/api/model/paginationPaciente.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Paciente } from './paciente';


export interface PaginationPaciente {
    total?: number;
    limit?: number;
    page?: number;
    docs?: Array<Paciente>;
}