// admin/src/modulos/api/model/especialidade.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EspecialidadeFamilia } from './especialidadeFamilia';
import { EspecialidadeGrupo } from './especialidadeGrupo';


export interface Especialidade {
    grupo?: EspecialidadeGrupo;
    familia?: EspecialidadeFamilia;
    /**
     * Código CBO da especialidade médica
     */
    codigo?: string;
    /**
     * descricao CBO da especialidade médica
     */
    descricao?: string;
    /**
     * descricao CBO da especialidade médica (redundância)
     */
    titulo?: string;
    sinonimos?: Array<string>;
}