// admin/src/modulos/api/model/pacienteDadosMedicos.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { PacienteDadosMedicosHistoricoMedico } from './pacienteDadosMedicosHistoricoMedico';
import { PacienteDadosMedicosTipoSanguineo } from './pacienteDadosMedicosTipoSanguineo';


export interface PacienteDadosMedicos {
    /**
     * Peso do paciente em Kg
     */
    peso?: number;
    /**
     * Altura do paciente em metros
     */
    altura?: number;
    tipoSanguineo?: PacienteDadosMedicosTipoSanguineo;
    /**
     * informar alergias, medicamentos, 1 por linha
     */
    alergiasMedicamentosas?: Array<string>;
    /**
     * informar outras alergias do paciente, 1 por linha
     */
    alergiasDiversas?: Array<string>;
    /**
     * informar outras alergias do paciente, 1 por linha
     */
    historicoMedico?: Array<PacienteDadosMedicosHistoricoMedico>;
}