// admin/src/modulos/api/model/pacienteFiliacao.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PacienteFiliacao {
    /**
     * Nome do mãe do paciente
     */
    mae?: string;
    /**
     * Nome do pai do paciente
     */
    pai?: string;
}