// admin/src/modulos/api/model/pacienteDadosMedicosTipoSanguineo.ts
/**
 * Receita Digital
 * Definição da API da Receita Digital 
 *
 * OpenAPI spec version: 1.0.6
 * Contact: fran.fig@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PacienteDadosMedicosTipoSanguineo {
    /**
     * Grupo sanguíneo do paciente
     */
    abo?: PacienteDadosMedicosTipoSanguineo.AboEnum;
    /**
     * Fator Rh (Rhesus)
     */
    rh?: PacienteDadosMedicosTipoSanguineo.RhEnum;
}
export namespace PacienteDadosMedicosTipoSanguineo {
    export type AboEnum = 'A' | 'B' | 'AB' | 'O' | '';
    export const AboEnum = {
        A: 'A' as AboEnum,
        B: 'B' as AboEnum,
        AB: 'AB' as AboEnum,
        O: 'O' as AboEnum,
        Empty: '' as AboEnum
    }
    export type RhEnum = '+' | '-' | '';
    export const RhEnum = {
        Plus: '+' as RhEnum,
        Minus: '-' as RhEnum,
        Empty: '' as RhEnum
    }
}